clang tidy

+++ src/network/SignalerMessageType.h
@@ -3,7 +3,7 @@
 namespace otherside
 {
 
-enum class MsgType : std::uint32_t
+enum class MsgType : std::uint8_t
 {
     PING,
     ACK_CONNECTION,
